# lib/api-zod/src/generated/types/cloudFile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * NovaOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CloudFile {
  id: number;
  workspaceId: string;
  path: string;
  language: string;
  content?: string;
  createdByUserId: string;
  lastEditedByUserId: string;
  version: number;
  deleted?: boolean;
  createdAt: Date;
  updatedAt: Date;
}